hex colors and move education to about folder
